Initialize repo
Subscribe to a tpic from Ubidots to turn on/off the LED on board. Also,
publish data converted from the ADC in a topic for Ubidots. Finally,
Add a license and a README file to explain how to indicate sensitive
data and to indicate the minimum requeriments of the project.